toolkit/mozapps/downloads/tests/chrome/test_cleanup_search.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * Test for bug 414850 to make sure only downloads that are shown when
 * searching are cleared and afterwards, the default list is shown.
 *
 * Test bug 430486 to make sure the Clear list button is disabled only when
 * there are no download items visible.
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

function test()
{
  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  let dm = Cc["@mozilla.org/download-manager;1"].
           getService(Ci.nsIDownloadManager);
  let db = dm.DBConnection;

  // Empty any old downloads
  db.executeSimpleSQL("DELETE FROM moz_downloads");

  // Make a file name for the downloads
  let file = Cc["@mozilla.org/file/directory_service;1"].
             getService(Ci.nsIProperties).get("TmpD", Ci.nsIFile);
  file.append("cleanUp");
  let filePath = Cc["@mozilla.org/network/io-service;1"].
                 getService(Ci.nsIIOService).newFileURI(file).spec;

  let stmt = db.createStatement(
    "INSERT INTO moz_downloads (name, target, source, state) " +
    "VALUES (?1, ?2, ?3, ?4)");

  try {
    for (let site of ["delete.me", "i.live"]) {
      stmt.bindByIndex(0, "Super Pimped Download");
      stmt.bindByIndex(1, filePath);
      stmt.bindByIndex(2, "http://" + site + "/file");
      stmt.bindByIndex(3, dm.DOWNLOAD_FINISHED);

      // Add it!
      stmt.execute();
    }
  }
  finally {
    stmt.reset();
    stmt.finalize();
  }

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testPhase = 0;
  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic != DLMGR_UI_DONE)
        return;

      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let $ = aId => win.document.getElementById(aId);
      let downloadView = $("downloadView");
      let searchbox = $("searchbox");
      let clearList = $("clearListButton");

      // The list must have built, so figure out what test to do
      switch (testPhase++) {
        case 0:
          // Make sure the button is initially enabled
          is(clearList.disabled, false, "Clear list is enabled for default 2 item view");

          // Search for multiple words in any order in all places
          searchbox.value = "delete me";
          searchbox.doCommand();

          break;
        case 1:
          // Search came back with 1 item
          is(downloadView.itemCount, 1, "Search found the item to delete");
          is(clearList.disabled, false, "Clear list is enabled for search matching 1 item");

          // Clear the list that has the single matched item
          clearList.doCommand();

          break;
        case 2:
          // Done rebuilding with one item left
          is(downloadView.itemCount, 1, "Clear list rebuilt the list with one");
          is(clearList.disabled, false, "Clear list still enabled for 1 item in default view");

          // Clear the whole list
          clearList.doCommand();

          break;
        case 3:
          // There's nothing left
          is(downloadView.itemCount, 0, "Clear list killed everything");
          is(clearList.disabled, true, "Clear list is disabled for no items");

          // We're done!
          win.close();
          obs.removeObserver(testObs, DLMGR_UI_DONE);
          SimpleTest.finish();

          break;
      }
    }
  };
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
